Split model module in uniffi to make it easier to maintain.
