+++ src/core/orchestrator/mod.rs
@@ -1,7 +1,7 @@
 use crate::{
     core::util::get_type_name,
     uniffi::{
-        model::PodJob,
+        model::pod::PodJob,
         orchestrator::{Orchestrator, PodRun},
     },
 };
