+++ src/core/pipeline.rs
@@ -1,4 +1,4 @@
-use crate::uniffi::model::Kernel;
+use crate::uniffi::model::pipeline::Kernel;
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Deserialize, Serialize)]
