+++ src/uniffi/model/pod.rs
@@ -1,35 +1,26 @@
 use crate::{
     core::{
-        crypto::{hash_blob, hash_buffer, make_random_hash},
-        graph::make_graph,
+        crypto::{hash_blob, hash_buffer},
         model::{
             deserialize_pod, deserialize_pod_job, serialize_hashmap, serialize_hashmap_option,
             to_yaml,
         },
-        pipeline::PipelineNode,
         validation::validate_packet,
     },
-    uniffi::{error::Result, orchestrator::Status},
+    uniffi::{
+        error::Result,
+        model::{
+            Annotation,
+            packet::{PathInfo, PathSet, URI},
+        },
+        orchestrator::Status,
+    },
 };
 use derive_more::Display;
 use getset::CloneGetters;
-use petgraph::graph::DiGraph;
 use serde::{Deserialize, Serialize};
 use std::{collections::HashMap, path::PathBuf, sync::Arc};
 use uniffi;
-
-/// Available models.
-#[derive(uniffi::Enum, Debug)]
-pub enum ModelType {
-    /// See [`Pod`].
-    Pod,
-    /// See [`PodJob`].
-    PodJob,
-    /// See [`PodResult`].
-    PodResult,
-}
-
-// --- core model structs ---
 
 /// A reusable, containerized computational unit.
 #[derive(
@@ -241,127 +232,6 @@ impl PodResult {
     }
 }
 
-/// Computational dependencies as a [DAG](https://en.wikipedia.org/wiki/Directed_acyclic_graph).
-#[derive(uniffi::Object, Debug, Display, CloneGetters, Clone, Deserialize, Serialize)]
-#[getset(get_clone, impl_attrs = "#[uniffi::export]")]
-#[display("{self:#?}")]
-#[uniffi::export(Display)]
-pub struct Pipeline {
-    /// Computational DAG in-memory.
-    #[getset(skip)]
-    pub graph: DiGraph<PipelineNode, ()>,
-    /// Exposed, internal input specification. Each input may be fed into more than one node/key if desired.
-    pub input_spec: HashMap<String, Vec<SpecURI>>,
-    /// Exposed, internal output specification. Each output is associated with only one node/key.
-    pub output_spec: HashMap<String, SpecURI>,
-}
-
-#[uniffi::export]
-impl Pipeline {
-    /// Construct a new pipeline instance.
-    ///
-    /// # Errors
-    ///
-    /// Will return `Err` if there is an issue initializing a `Pipeline` instance.
-    #[uniffi::constructor]
-    pub fn new(
-        graph_dot: &str,
-        metadata: HashMap<String, Kernel>,
-        input_spec: &HashMap<String, Vec<SpecURI>>,
-        output_spec: &HashMap<String, SpecURI>,
-    ) -> Result<Self> {
-        let graph = make_graph(graph_dot, metadata)?;
-        Ok(Self {
-            graph,
-            input_spec: input_spec.clone(),
-            output_spec: output_spec.clone(),
-        })
-    }
-}
-
-/// A compute pipeline job that supplies input/output targets.
-#[expect(
-    clippy::field_scoped_visibility_modifiers,
-    reason = "Temporary until a proper hash is implemented."
-)]
-#[derive(uniffi::Object, Debug, Display, CloneGetters, Deserialize, Serialize, Clone)]
-#[getset(get_clone, impl_attrs = "#[uniffi::export]")]
-#[display("{self:#?}")]
-#[uniffi::export(Display)]
-pub struct PipelineJob {
-    /// todo: replace with a consistent hash
-    #[getset(skip)]
-    pub(crate) hash: String,
-    /// A pipeline to base the pipeline job on.
-    pub pipeline: Arc<Pipeline>,
-    /// Attached, external input packet. Applies cartesian product by default on keys pointing to the same node.
-    pub input_packet: HashMap<String, Vec<PathSet>>,
-    /// Attached, external output directory.
-    pub output_dir: URI,
-}
-
-#[expect(clippy::excessive_nesting, reason = "Nesting manageable.")]
-#[uniffi::export]
-impl PipelineJob {
-    /// Construct a new pipeline job instance.
-    ///
-    /// # Errors
-    ///
-    /// Will return `Err` if there is an issue initializing a `PipelineJob` instance.
-    #[uniffi::constructor]
-    pub fn new(
-        pipeline: Arc<Pipeline>,
-        input_packet: &HashMap<String, Vec<PathSet>>,
-        output_dir: &URI,
-        namespace_lookup: &HashMap<String, PathBuf>,
-    ) -> Result<Self> {
-        validate_packet("input".into(), &pipeline.input_spec, input_packet)?;
-        let input_packet_with_checksum = input_packet
-            .iter()
-            .map(|(path_set_key, path_sets)| {
-                Ok((
-                    path_set_key.clone(),
-                    path_sets
-                        .iter()
-                        .map(|path_set| {
-                            Ok(match path_set {
-                                PathSet::Unary(blob) => {
-                                    PathSet::Unary(hash_blob(namespace_lookup, blob)?)
-                                }
-                                PathSet::Collection(blobs) => PathSet::Collection(
-                                    blobs
-                                        .iter()
-                                        .map(|blob| hash_blob(namespace_lookup, blob))
-                                        .collect::<Result<_>>()?,
-                                ),
-                            })
-                        })
-                        .collect::<Result<_>>()?,
-                ))
-            })
-            .collect::<Result<_>>()?;
-
-        Ok(Self {
-            hash: make_random_hash(),
-            pipeline,
-            input_packet: input_packet_with_checksum,
-            output_dir: output_dir.clone(),
-        })
-    }
-}
-
-// --- util types ---
-
-/// Standard metadata structure for all model instances.
-#[derive(uniffi::Record, Serialize, Deserialize, Debug, PartialEq, Eq, Clone)]
-pub struct Annotation {
-    /// A unique name.
-    pub name: String,
-    /// A unique semantic version.
-    pub version: String,
-    /// A long form description.
-    pub description: String,
-}
 /// Specification for GPU requirements in computation.
 #[derive(uniffi::Record, Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
 pub struct GPURequirement {
@@ -372,6 +242,7 @@ pub struct GPURequirement {
     /// Number of GPU cards required.
     pub count: u16,
 }
+
 /// GPU model specification.
 #[derive(uniffi::Enum, Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
 pub enum GPUModel {
@@ -380,81 +251,3 @@ pub enum GPUModel {
     /// AMD-manufactured card where `String` is the specific model e.g. ???
     AMD(String),
 }
-/// Streams are named and represent an abstraction for the file(s) that represent some particular
-/// data.
-#[derive(uniffi::Record, Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
-pub struct PathInfo {
-    /// Path to stream file or directory.
-    pub path: PathBuf,
-    /// Naming pattern for the stream.
-    pub match_pattern: String,
-}
-/// A single BLOB or a collection of BLOBs.
-#[derive(uniffi::Enum, Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
-#[serde(untagged)]
-pub enum PathSet {
-    /// A single BLOB.
-    Unary(Blob),
-    /// A series of BLOBs.
-    Collection(Vec<Blob>),
-}
-/// Location of BLOB data.
-#[derive(uniffi::Record, Serialize, Deserialize, Debug, Clone, PartialEq, Eq, Default)]
-pub struct URI {
-    /// Namespace alias.
-    pub namespace: String,
-    /// Path within namespace.
-    pub path: PathBuf,
-}
-
-/// BLOB with metadata.
-#[derive(uniffi::Record, Serialize, Deserialize, Debug, Clone, PartialEq, Eq, Default)]
-pub struct Blob {
-    /// BLOB available options.
-    pub kind: BlobKind,
-    /// BLOB location.
-    pub location: URI,
-    /// BLOB contents checksum.
-    pub checksum: String,
-}
-/// File or directory options for BLOBs.
-#[derive(uniffi::Enum, Serialize, Deserialize, Debug, Clone, PartialEq, Eq, Default)]
-pub enum BlobKind {
-    /// A single file.
-    #[default]
-    File,
-    /// A single directory.
-    Directory,
-}
-/// A node in a computational pipeline.
-#[derive(uniffi::Enum, Debug, Clone, Deserialize, Serialize)]
-pub enum Kernel {
-    /// Pod reference.
-    Pod {
-        /// See [`Pod`].
-        r#ref: Arc<Pod>,
-    },
-    /// Cartesian product operation. See [`crate::core::operator::JoinOperator`].
-    JoinOperator,
-    /// Rename a path set key operation.
-    MapOperator {
-        /// See [`crate::core::operator::MapOperator`].
-        map: HashMap<String, String>,
-    },
-}
-/// Index from pipeline node into pod specification.
-#[derive(uniffi::Record, Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
-pub struct SpecURI {
-    /// Node reference name in pipeline.
-    pub node: String,
-    /// Specification key.
-    pub key: String,
-}
-
-// --- utils ----
-
-uniffi::custom_type!(PathBuf, String, {
-    remote,
-    try_lift: |val| Ok(PathBuf::from(&val)),
-    lower: |obj| obj.display().to_string(),
-});
